sequence_predict/requirements.txt
numpy==1.17.4
Keras==2.3.1
matplotlib==3.1.1
requests==2.22.0
keras_tcn==2.8.3
xgboost==0.90
tensorflow_gpu==2.0.0
tqdm==4.36.1
pandas==0.25.1
psutil==5.6.3
GPUtil==1.4.0
scikit_learn==0.22.1
tensorflow==2.1.0
torch==1.4.0
